# Store Turns per Intersection, instead of a massive BTreeMap. #368