--- a/game/src/challenges/prebake.rs
+++ b/game/src/challenges/prebake.rs
@@ -31,7 +31,7 @@ pub fn prebake_all() {
         //MapName::seattle("phinney"),
         MapName::seattle("qa"),
         MapName::seattle("rainier_valley"),
-        MapName::seattle("wallingford"),
+        //MapName::seattle("wallingford"),
     ] {
         let map = map_model::Map::load_synchronously(name.path(), &mut timer);
         let scenario: Scenario =