… data/MANIFEST.json