Minimal file size savings, but substantial runtime improvement!

See https://github.com/a-b-street/abstreet/issues/368#issuecomment-849891303 for context on this idea.

# Performance impact

I ran montlake in headless mode a few times:

- before: 17.8s, 20.6s, 21.3s, 17.9s
- after: 13.4s, 13.1s, 15.5s, 14.0s

Quite a bit of variance there, but consistently this new approach is faster.

More dramatically... a single run of `game --prebake` drops from 12m36s to 9m11s!

# File size impact

Not very much -- lakeslice from 25MB to 24MB. All we're saving is all of the `TurnIDs` being listed twice, from the old `BTreeMap`.

## Idea: bitpacking

I was curious if we could squeeze down `TurnID` when serializing from 3 u32's (the intersection, src lane, dst lane). In huge_seattle, we have about 265k lanes and 41k intersections. 19 bits could fit about 1 million lane IDs. We'd only need 16 bits for intersections, giving us a max of 65k. But that's 54 bits total. So instead of 32*3=96 bits per TurnID, we could fit it in 64 bits. It could be a future optimization just at the serde level, but I don't think it's worth it in the meantime.

# API impact

In the issue, we had talked about a more dramatic change: making `TurnID` encode `IntersectionID` and then some index into the list of turns there. That would give us even faster runtime performance; two array lookups, instead of an array lookup and a linear scan. And it would definitely let us compress `TurnID` to a single `u32`.

But that change would be very invasive to the code base; loads of places currently wind up with a `TurnID` one way or another, and conveniently look up its 3 fields. So as future work, we could try out this change, maybe with some careful layers of translation in `map_model` to hand out the old `TurnIDs`, maybe with an added index. But external code could still grab `src` and `dst` quickly.

# TODO

I still need to regenerate all maps because of this format change. I'm going to try to bundle that with another small importer change today to save on cloud costs.