@dabreegster 2021-06-11
Thanks for the review! Also sticking a small stop sign precedence fix in this PR, so I can easily regenerate maps with both changes. Kicking that off now